scripts/requirements_chatbot.txt
# PYTHON 3.10 ONLY

numpy==1.26.4
torch==2.1.2
sentence-transformers==2.6.1

faiss-cpu==1.7.4

langchain==0.1.20
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-huggingface==0.0.3

python-dotenv==1.0.1
tqdm
